+++ packages/errors-nest/src/errors/index.ts
@@ -2,6 +2,7 @@ export * from './auth.errors.js';
 export * from './bridge.errors.js';
 export * from './health-check.errors.js';
 export * from './event.errors.js';
+export * from './post.errors.js';
 export * from './database.errors.js';
 export * from './common.errors.js';
 export * from './social.errors.js';